+++ app/src/js/components/Granules/index.js
@@ -1,5 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import PropTypes from 'prop-types';
+import { Helmet } from 'react-helmet';
 import { get } from 'object-path';
 import { connect } from 'react-redux';
 import { withRouter, Redirect, Route, Switch } from 'react-router-dom';
@@ -36,6 +37,9 @@ const Granules = ({
 
   return (
     <div className='page__granules'>
+      <Helmet>
+        <title> Granules </title>
+      </Helmet>
       <DatePickerHeader onChange={query} heading={strings.granules}/>
       <div className='page__content'>
         <div className='wrapper__sidebar'>