+++ app/src/js/components/Pdr/index.js
@@ -1,5 +1,6 @@
 'use strict';
 import React, { useState, useEffect } from 'react';
+import { Helmet } from 'react-helmet';
 import PropTypes from 'prop-types';
 import { get } from 'object-path';
 import { connect } from 'react-redux';
@@ -37,6 +38,9 @@ const Pdrs = ({
 
   return (
     <div className='page__pdrs'>
+      <Helmet>
+        <title> Cumulus PDRs </title>
+      </Helmet>
       <DatePickerHeader onChange={query} heading={strings.pdrs}/>
       <div className='page__content'>
         <div className='wrapper__sidebar'>