+++ app/src/js/components/Providers/overview.js
@@ -1,6 +1,6 @@
 'use strict';
-
 import React from 'react';
+import { Helmet } from 'react-helmet';
 import { connect } from 'react-redux';
 import { withRouter, Link } from 'react-router-dom';
 import { get } from 'object-path';
@@ -76,6 +76,9 @@ class ProvidersOverview extends React.Component {
     const overview = this.renderOverview(providerStatus);
     return (
       <div className='page__component'>
+        <Helmet>
+          <title> Provider </title>
+        </Helmet>
         <section className='page__section page__section__header-wrapper'>
           <h1 className='heading--large heading--shared-content with-description'>Provider Overview</h1>
           {lastUpdated(queriedAt)}